aded middleware to check if is active user

+++ app/Services/AuthService.php
@@ -28,6 +28,10 @@ class AuthService
       return null;
     }
 
+    if (!$user->is_active) {
+      abort(403, 'هذا الحساب معطل، يرجى التواصل مع الإدارة.');
+    }
+
     $user->assignRole('customer');
     $user->tokens()->delete();
     $token = $user->createToken('auth_token')->plainTextToken;
